requirements.txt
absl-py==2.3.0
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.6.15
click==8.2.1
cmdstanpy==1.2.5
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.115.13
fastapi-cli==0.0.7
fonttools==4.58.4
h11==0.16.0
holidays==0.75
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
immutabledict==4.2.1
importlib_resources==6.5.2
Jinja2==3.1.6
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
numpy==2.2.6
ortools==9.14.6206
packaging==25.0
pandas==2.3.0
pillow==11.2.1
prophet==1.1.7
protobuf==6.31.1
psycopg2==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
rich==14.0.0
rich-toolkit==0.14.7
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
stanio==0.5.1
starlette==0.46.2
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uvicorn==0.34.3
watchfiles==1.1.0
websockets==15.0.1
